+++ libs/zhi-xmlrpc-middleware/CHANGELOG.md
@@ -1,5 +1,20 @@
 # zhi-xmlrpc-middleware
 
+## 0.6.5
+
+### Patch Changes
+
+- feat: add node cmd
+- Updated dependencies
+  - zhi-device@2.8.1
+
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.8.0
+
 ## 0.6.3
 
 ### Patch Changes